code/requirements.txt
Flask==2.1.1
pycryptodome==3.14.1
pyOpenSSL==22.0.0